reorganized under examples
